+++ src/Skybrud.Social/Instagram/Objects/InstagramLocation.cs
@@ -13,7 +13,7 @@ namespace Skybrud.Social.Instagram.Objects {
         /// <summary>
         /// Gets the ID of the location.
         /// </summary>
-        public int Id { get; private set; }
+        public long Id { get; private set; }
 
         /// <summary>
         /// The name of the location.
@@ -65,7 +65,7 @@ namespace Skybrud.Social.Instagram.Objects {
         public static InstagramLocation Parse(JsonObject obj) {
             if (obj == null) return null;
             return new InstagramLocation(obj) {
-                Id = obj.GetInt32("id"),
+                Id = obj.GetInt64("id"),
                 Name = obj.GetString("name"),
                 Latitude = obj.GetDouble("latitude"),
                 Longitude = obj.GetDouble("longitude")
